+++ lib/riverpod_paginated_list.dart
@@ -4,6 +4,7 @@
 /// external items injection, and support for both zero-based and one-based pagination.
 library;
 
+export 'src/paginated_list_builder.dart';
 export 'src/paginated_list_config.dart';
 export 'src/paging.dart';
 export 'src/skeleton_config.dart';